LayoutTests/platform/win/fast/forms/reset-autofilled-expected.txt

This tests that a field's autofilled state is reverted when the form is reset. It can only be run using DumpRenderTree.

FAIL countAutofilledElements() should be 1. Was 0.
PASS countAutofilledElements() is 0
